waveform: drop the hand-rolled and plotters implementations
Both implementations are superseded by the Waveform builder, so remove
them along with the Channels and ChannelInterleavement enums that only
existed to describe their i16 input. Splitting interleaved stereo data
is now the job of deinterleave_stereo(); the LLRR layout is dropped
since LRLR is the de-facto standard.

Migrate the remaining callers: the sine test utility produces f32
amplitudes in [-1.0, 1.0] instead of scaled i16 values, and decode_mp3()
does the same. The MP3 test that lived in the deleted png_file.rs moves
to src/tests/ next to the other visual tests.

+++ src/tests/testutil/mod.rs
@@ -15,8 +15,8 @@ pub const TEST_SAMPLES_DIR: &str = "test/samples";
 /// If tests create files, they should be stored here.
 pub const TEST_OUT_DIR: &str = "target/test_out";
 
-/// Returns an MP3 as decoded i16 samples and with LRLR interleavement.
-pub fn decode_mp3(file: &Path) -> Vec<i16> {
+/// Returns an MP3 as decoded f32 samples and with LRLR interleavement.
+pub fn decode_mp3(file: &Path) -> Vec<f32> {
     let file = File::open(file).unwrap();
     let mss = MediaSourceStream::new(Box::new(file), Default::default());
     let probed = get_probe()
